rs: fix trace re-install stacking, matcher precision, match-token form
Addresses the four findings from the automated review of the Rust port.
Three were real defects in this port and are fixed; two are faithful
reproductions of canonical TypeScript behaviour and are recorded instead
of silently diverged.

Fixed:

1. **Re-installing the plugin stacked the trace.** The engine ACCUMULATES
   subscribers and parse-prepare hooks, so a second `apply` registered a
   second set: duplicate banners and duplicate events, and a later,
   narrower selection could not switch the first set off. `derive`
   re-runs a parent's plugins on the child, so this was reachable without
   anyone calling `apply` twice by hand.

   `trace::install` now registers its callbacks exactly once per instance
   and keeps the live selection in a `debug.trace` decoration that later
   installs replace. It is called for `None` too, so re-applying without
   trace turns tracing OFF rather than leaving the old streams running.

2. **`DebugLexMatcher::order` truncated fractional priorities.** It was
   `i64` over the engine's `f64`, so 1.2 and 1.8 both reported as 1. It
   is now `f64` — which also matches the canonical TypeScript field (a
   plain JavaScript number) more closely than Go's `int` can.

3. **A function-backed match token emitted invalid ABNF.** The legend got
   `T = ; <match function …>`, and since `;` runs to end of line that
   defines a rule with no elements. This port invented that form: the
   canonical runtime special-cases only a RegExp and otherwise falls
   through to the built-in description, so it now does the same and emits
   `<built-in NAME>`.

Not fixed here, and why — both verified against `ts/src/debug.ts` and now
recorded in `docs/reference.md` § "Known limitations inherited from the
canonical runtime":

- A token whose bare name matches a rule name collides in `abnf()`,
  producing a self-referential `NR = NR` plus a duplicate definition.
- An unrecognised match regex emits a legend entry that is only a comment
  (`T = ; /…/`), leaving a rule with no elements.

Both are genuine bugs, and both are present identically in the canonical
TypeScript. The repository's authority rule is that a port exposing a TS
defect means fixing TS first and pinning the corrected behaviour in a
shared fixture; changing Rust alone would make the emitted ABNF diverge
silently from the runtime that defines it. Neither case is covered by a
fixture today, so adding one is part of that TypeScript change.

Every fix carries a regression test: the `reapplying_*` trio,
`deriving_a_child_does_not_stack_trace_subscribers`,
`lexer_matcher_order_keeps_fractional_priorities`, and
`a_function_backed_match_token_gets_a_valid_abnf_form`. The
re-install test compares whole event counts, not just banners, so a
stacked second set fails it.

Verified: `cargo test --all-targets` (37 tests, up from 31), `cargo
clippy --all-targets --all-features -- -D warnings`, `cargo fmt --check`,
and `GOWORK=off go test ./...` still green.

+++ rs/src/model.rs
@@ -124,10 +124,16 @@ pub struct DebugRuleEdges {
 /// instead — the same limit the Go port documents. `make` is the Go
 /// analogue of the TypeScript factory name and has no Rust counterpart
 /// (function values carry no name), so it is always empty.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
+#[derive(Debug, Clone, PartialEq, Serialize)]
 pub struct DebugLexMatcher {
     /// Priority (lower runs first).
-    pub order: i64,
+    ///
+    /// Kept as `f64`, the engine's own type: priorities are not
+    /// necessarily whole, and truncating would report `1.2` and `1.8` as
+    /// the same matcher order. The canonical TypeScript field is a plain
+    /// JavaScript number, so this matches it; the Go port's `int` is a Go
+    /// engine limit.
+    pub order: f64,
     /// Registered matcher name.
     pub matcher: String,
     /// Matcher function name, when recoverable.
@@ -371,7 +377,7 @@ pub(crate) fn lexer(parser: &Tabnas) -> Vec<DebugLexMatcher> {
     matchers
         .into_iter()
         .map(|matcher| DebugLexMatcher {
-            order: matcher.order as i64,
+            order: matcher.order,
             matcher: matcher.name.clone(),
             make: String::new(),
         })
